Fix issue with Delete reloading top of Page
In my experience, clicking the delete button on a bit tag, caused the page to reload at the top because it's an anchor tag with an href. This probably shouldn't be an anchor tag, but if it is going to be, we don't want the browser interpreting that click event. So making the `href= "javascript:void(0);"` will stop that. You could also put an `e.stop()` function on the click.